requirements.txt
pyvmomi>=8.0.0
paramiko>=3.4.0
psutil>=5.9.0
psycopg2-binary>=2.9.9